+++ app/src/main/java/com/fintech/ozmaden_developerslife/ui/PostViewModel.kt
@@ -58,4 +58,19 @@ abstract class PostViewModel : ViewModel() {
     }
 
     protected abstract suspend fun loadNewPost()
+
+//    protected open suspend fun loadNewPost() {
+//        viewModelScope.launch {
+//            val rndPost = postRepository.getRandomPost()
+//            if (rndPost != null) {
+//                _post.postValue(rndPost!!)
+//                _description.postValue(rndPost.description)
+//                postHistory.add(rndPost)
+//                position++
+//                Log.d("RandomViewModel", "Got random post and inserted it!")
+//            } else {
+//                Log.d("RandomViewModel", "Could not insert post!")
+//            }
+//        }
+//    }
 }
